+++ unit_tests/utils/setup_project.sh
@@ -32,21 +32,11 @@ pip install -e "$sd_root_dir/"
 # Make it easier to verify what was installed when developing this script.
 pip freeze > installed_packages.txt
 
-# Add simple_deploy to INSTALLED_APPS.
-sed -i "" "s/# Third party apps./# Third party apps.\n    'simple_deploy',/" blog/settings.py
-
 # Make an initial commit.
 git init
 git add .
 git commit -am "Initial commit."
 git tag -am '' INITIAL_STATE
 
-
-# At this point simple_deploy expects `heroku create` to have been run,
-#   or expects us to use `--automate-all` to call `heroku create`.
-# How mock this call? Need to get output for `heroku apps:info`.
-
-# Run configuration-only version of simple_deploy.
-# DEV: Trying to move this call to conftest, in order to test invalid
-#   versions of this command.
-# python manage.py simple_deploy --unit-testing --platform heroku
+# Add simple_deploy to INSTALLED_APPS.
+sed -i "" "s/# Third party apps./# Third party apps.\n    'simple_deploy',/" blog/settings.py